Modificando tipos de variables para q sean todas unsigned long int (por si las moscas) ... y arreglando un error que tenía (gracias Gaby por avisarme)..era una pavada al final ..tenía dos veces definida la misma variables ..y en un lado la seteaba en cero jaja...q sopencaaaaa

+++ trunk/src/business/stego/BMPStego.cpp
@@ -5,7 +5,7 @@ BMPStego::BMPStego(std::string filename):LSBStegoBusiness(filename)
 
 }
 
-unsigned int BMPStego::getFreeSpace(){
+unsigned long int BMPStego::getFreeSpace(){
  if((imagen.getBpp()<=8)&&(!imagen.isGrayScale())){
    if(imagen.getPaletteSize()>16){ 
    	       palette.sortPaletteByDistance();
