+++ trunk/src/business/stego/ImagePalette.h
@@ -12,26 +12,26 @@
 class ImagePalette
 {
 	private:
-
+	
     	/*Representacion de la imagen*/
    		FIBitmap &imagen;
    	  	std::vector<unsigned int> new_palette_indexes;
 		unsigned int getRGBPos(unsigned int pos);
-
+		
 	public:
-
+	
 		ImagePalette();
 		ImagePalette(FIBitmap &imagen);
-
+	
 		unsigned int getNewPaletteIndex(unsigned int index);
 		void sortPaletteByDistance();
-		unsigned int doIndexesLSB(Pixel& pixel,const char* mensaje, long int size);
-		unsigned int doPaletteLSB(unsigned int first_bit,const char* mensaje, long int size);
-		std::string getMessageFromIndexes(Pixel& pixel,unsigned int longitud);
-		std::string getMessageFromPalette(unsigned int first,unsigned int longitud);
-
+		unsigned int doIndexesLSB(Pixel& pixel,const char* mensaje,unsigned long int size);		 
+		unsigned int doPaletteLSB(unsigned int first_bit,const char* mensaje,unsigned long int size);
+		std::string getMessageFromIndexes(Pixel& pixel,unsigned long int longitud);
+		std::string getMessageFromPalette(unsigned int first,unsigned long int longitud);
+	
 		virtual ~ImagePalette();
-
+		
 };
 
 #endif /*IMAGEPALETTE_H_*/
