+++ trunk/src/business/stego/StegoBusiness.h
@@ -16,18 +16,18 @@ protected:
 
 public:
     StegoBusiness();
-
+  
 	StegoBusiness(std::string filename);
-
+	
 	/*A implementar por las clases hijas: */
-	virtual unsigned int setMessage(unsigned long int first_bit,const char* mensaje, long int size)=0;
-	virtual std::string getMessage(unsigned long int first_bit,unsigned int longitud)=0;
-    virtual unsigned int getFreeSpace()=0;
+	virtual unsigned long int setMessage(unsigned long int first_bit,const char* mensaje,unsigned long int size)=0;
+	virtual std::string getMessage(unsigned long int first_bit,unsigned long int longitud)=0;
+    virtual unsigned long int getFreeSpace()=0;
     virtual unsigned int getFirstFreeBit()=0;
-
+    
 	std::string& getFilename();
 	void setFilename(std::string file);
-
+			
 	virtual ~StegoBusiness();
 };
 #endif /*STEGOBUSINESS_H_*/
